138: Allow for retrieving member from history (#143)
