+++ src/stories/history/history.ts
@@ -1,7 +1,10 @@
 import ShortcutResource from '@sx/base-resource'
+import Member from '@sx/members/member'
+import MembersService from '@sx/members/members-service'
 import HistoryActionInterface from '@sx/stories/history/actions/contracts/history-action-interface'
 import HistoryAction from '@sx/stories/history/actions/history-action'
 import HistoryInterface from '@sx/stories/history/contracts/history-interface'
+import {getHeaders} from '@sx/utils/headers'
 import UUID from '@sx/utils/uuid'
 
 
@@ -15,6 +18,14 @@ class History extends ShortcutResource<HistoryInterface> implements HistoryInter
 
   private instantiateActions() {
     this.actions = this.actions?.map(action => new HistoryAction(action))
+  }
+
+  /**
+   * Get an instance of the member that made the change
+   */
+  get member(): Promise<Member>{
+    const service = new MembersService({headers: getHeaders()})
+    return service.get(this.memberId)
   }
 
   actions: HistoryActionInterface[] | HistoryAction[]
