+++ src/utils/is-valid-datetime-format.ts
@@ -1,4 +1,6 @@
-export default function isValidDatetimeFormat(text: string): boolean {
-  const pattern = /^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$/
+function isValidDatetimeFormat(text: string): boolean {
+  const pattern = /^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d{1,3})?Z$/
   return pattern.test(text)
 }
+
+export default isValidDatetimeFormat
